devices/omron-env/requirements.txt
python-dotenv==0.19.0
paho-mqtt==1.5.1
bluepy==1.3.0
